Get/SetKernelStatus now go through `KernelAttribute`s of nest module
